Move LoadApplicationObjectFailedError up
