+++ lib/graphql.rb
@@ -106,3 +106,4 @@ require "graphql/deprecated_dsl"
 require "graphql/authorization"
 require "graphql/unauthorized_error"
 require "graphql/unauthorized_field_error"
+require "graphql/load_application_object_failed_error"
